Fixed names of files & seperated them into seperate files
