+++ apps/teamsjs-diagnostic-app/.hintrc
