+++ apps/teamsjs-diagnostic-app/api/local.settings.json
