+++ apps/teamsjs-diagnostic-app/appPackage/manifest.json
