+++ apps/teamsjs-diagnostic-app/src/apis/AppInstallDialogApi.tsx
