+++ apps/teamsjs-diagnostic-app/src/apis/BarCodeApi.tsx
