+++ apps/teamsjs-diagnostic-app/src/components/sample/AppInitializationScenario.css
@@ -1,0 +1,63 @@
+  .scenario-container {
+    display: flex;
+    gap: 20px;
+    margin-bottom: 20px;
+  }
+  
+  .scenario1-box {
+    width: 400px;
+    height: 400px;
+    background-color: lightgray;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    overflow: hidden;
+  }
+  
+  
+  
+  .scenario1-button {
+    padding: 10px 20px;
+    font-size: 16px;
+    cursor: pointer;
+    background-color: #464eb8;
+    color: white;
+  }
+  
+  .api-section {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 20px;
+    flex: 1;
+  }
+  
+  .api-header {
+    text-align: center;
+    margin-bottom: 10px;
+    font-weight: bold;
+  }
+  
+  .vertical-box-container {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    overflow-y: auto;
+    max-height: calc(100% - 40px);
+  }
+  
+  .vertical-box {
+    width: 100%;
+    background-color: white;
+    display: flex;
+    align-items: center;
+    border: 1px solid #ccc;
+    padding: 10px;
+    border-radius: 4px;
+  }
+  
+  .box-title {
+    flex-grow: 1;
+  }
+
