+++ apps/teamsjs-diagnostic-app/src/components/sample/AppInitializationScenario.tsx
@@ -1,0 +1,66 @@
+import React, { useEffect } from 'react';
+import './AppInitializationScenario.css'; // Import CSS file for AppInitialization
+
+import { app } from '@microsoft/teams-js';
+import { registerOnResume } from '../../apis/AppApi';
+import { authenticateUser } from '../../apis/AuthenticationStart';
+
+const AppInitializationScenario: React.FC = () => {
+  useEffect(() => {
+    app.initialize();
+  }, []);
+
+  const runAppInitializationScenario = async () => {
+    try {
+      console.log('Running App Initialization Scenario...');
+      console.log('Attempting to register on resume handler...');
+      await registerOnResume();
+      console.log('Attempting to authenticate user...');
+      const authSuccess = await authenticateUser();
+      if (authSuccess) {
+        console.log('App Initialization Scenario successfully completed');
+      } else {
+        console.log('User not authenticated');
+        showSignInPopup();
+      }
+    } catch (error: any) {
+      console.log(`App initialization scenario failed. ${error.message}`);
+    }
+  };
+
+  const showSignInPopup = () => {
+    console.log('Showing sign-in popup...');
+  };
+
+  return (
+    <div>
+      <h2>App Initialization Scenario</h2>
+      <p>Click the button to run the app initialization scenario.</p>
+      <div className="scenario-container">
+        <div className="scenario1-box">
+          <button
+            className="scenario1-button"
+            onClick={runAppInitializationScenario}
+            type="button"
+            data-testid="run-scenario-button"
+          >
+            Run Scenario
+          </button>
+          <div className="api-section">
+            <div className="api-header">APIs Being Run:</div>
+            <div className="vertical-box-container">
+              <div className="vertical-box">
+                <span className="box-title">1. app</span>
+              </div>
+              <div className="vertical-box">
+                <span className="box-title">2. authentication</span>
+              </div>
+            </div>
+          </div>
+        </div>
+        </div>
+    </div>
+  );
+};
+
+export default AppInitializationScenario;
