+++ apps/teamsjs-diagnostic-app/src/components/sample/CustomScenario.css
@@ -1,0 +1,107 @@
+.sr-only {
+    border: 0;
+    clip: rect(0, 0, 0, 0);
+    height: 1px;
+    margin: -1px;
+    overflow: hidden;
+    padding: 0;
+    position: absolute;
+    width: 1px;
+  }
+  
+  .custom-scenario-box {
+    width: 400px;
+    height: 400px;
+    background-color: lightgray;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 20px;
+    position: relative;
+  }
+  
+  .all-api-container {
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    align-items: flex-start;
+    overflow-y: auto;
+    max-height: 600px;
+    background-color: lightgray;
+    padding: 20px;
+    height: 400px;
+    margin-top: 88px;
+  }
+  
+  .all-api-box {
+    width: 400px;
+    height: 400px;
+    overflow: auto;
+  }  
+  
+  
+  .set-scenario-button {
+    font-size: 16px;
+    cursor: pointer;
+    position: absolute;
+    top: 10px;
+    right: 10px;
+  }
+  
+  .api-section {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin-top: 20px;
+    flex: 1;
+  }
+  
+  .api-header {
+    text-align: center;
+    margin-bottom: 10px;
+    font-weight: bold;
+  }
+  
+  .vertical-box-container {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    overflow-y: auto;
+    max-height: calc(100% - 40px);
+  }
+  
+  .vertical-box {
+    width: 100%;
+    background-color: white;
+    display: flex;
+    align-items: center;
+    border: 1px solid #ccc;
+    padding: 10px;
+    border-radius: 4px;
+  }
+  
+  .box-dropdown {
+    margin-left: auto;
+  }
+  
+  .box-title {
+    flex-grow: 1;
+  }
+  
+  .box-dropdown {
+    margin-left: 10px;
+  }
+  
+  
+  .scenario2-header {
+    gap: 20px;
+  }
+  
+  
+  .search-bar {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 10px;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+  }
