+++ apps/teamsjs-diagnostic-app/src/components/sample/LogConsole.css
@@ -1,0 +1,30 @@
+
+  .scenario-container {
+    display: flex;
+    gap: 20px;
+    margin-bottom: 20px;
+  }
+  
+  .horizontal-box {
+    width: 1300px;
+    height: 600px;
+    overflow: auto;
+    background-color: lightgray;
+  }
+  
+  .content {
+    width: 100%;
+    height: 100%;
+  }
+  
+  .log-console {
+    margin-top: 20px;
+    padding: 20px;
+    background-color: lightgray;
+  }
+  
+  .log-statement {
+    padding: 5px;
+    border-bottom: 1px solid #ccc;
+    overflow: auto;
+  }
