+++ apps/teamsjs-diagnostic-app/src/components/sample/Scenarios.css
@@ -1,0 +1,24 @@
+.scenarios-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .scenarios-row {
+    display: flex;
+    gap: 20px;
+    margin-bottom: 20px;
+  }
+  
+  .scenario-column {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .custom-and-apis-container {
+    display: flex;
+    flex-direction: column;
+    gap: 20px;
+  }
+
